% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_data.R
\name{conta_normal}
\alias{conta_normal}
\title{Sample contaminated normal}
\usage{
conta_normal(
  n,
  Omega,
  byrow = FALSE,
  cont_rate = 0.05,
  mu = 10,
  sd = sqrt(0.2)
)
}
\arguments{
\item{n}{samplesize}

\item{Omega}{precision matrix of the normal}

\item{byrow}{whether the contamination happened by row? FALSE stand for cellwise contamination}

\item{cont_rate}{how many cells/rows are contaminated?}

\item{mu}{mean of the contamination}

\item{sd}{standard deviation of the contamination}
}
\value{
a matrix of contaminated (multivariate) normal distributed data, row as sample
}
\description{
This function samples normal distribution with normal contamination
}
