% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corSpearman}
\alias{corSpearman}
\title{Spearman correlation}
\usage{
corSpearman(data)
}
\arguments{
\item{data}{the n by p raw data matrix}
}
\value{
a matrix with dimension p by p of spearman correlations
}
\description{
This routine calculates the Spearman correlation
}
\examples{
corSpearman(matrix(rnorm(500),100,5)) 
}
