% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corSpearmanmat}
\alias{corSpearmanmat}
\title{Spearman correlation}
\usage{
corSpearmanmat(data)
}
\arguments{
\item{data}{the n by p raw data matrix}
}
\value{
a matrix with dimension p by p of spearman correlations
}
\description{
This routine calculates the Spearman correlation
}
\examples{
corSpearmanmat(matrix(rnorm(500),100,5)) 
}
