% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covGKmat}
\alias{covGKmat}
\title{Gnanadesikan-Kettenring estimator for *covariance*}
\usage{
covGKmat(data)
}
\arguments{
\item{data}{the n by p raw data matrix}
}
\value{
a matrix with dimension p by p, GK estimator, note that it's not necessarily positive
}
\description{
This routine calculates the Gnanadesikan-Kettenring estimator, diagonal will be MAD
}
