% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covOGKmat}
\alias{covOGKmat}
\title{Orthogonalized Gnanadesikan–Kettenring (OGK) estimator for *covariance*}
\arguments{
\item{data}{the n by p raw data matrix}
}
\value{
a matrix with dimension p by p, OGK estimator
}
\description{
This routine calculates the Orthogonalized Gnanadesikan–Kettenring (OGK) estimator for *covariance*, using scale estimation of Gn, as in Maronna and Zamar
}
