% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covSpearmanU}
\alias{covSpearmanU}
\title{SpearmanU estimator for *covariance*}
\usage{
covSpearmanU(data)
}
\arguments{
\item{data}{the n by p raw data matrix}
}
\value{
a matrix with dimension p by p of spearmanU correlation
}
\description{
This routine calculates the SpearmanU, the pairwise covariance matrix estimator proposed in Oellererand Croux
}
\examples{
covSpearmanU(matrix(rnorm(500),100,5)) 
}
