% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covSpearmanUmat}
\alias{covSpearmanUmat}
\title{SpearmanU estimator for *covariance*}
\usage{
covSpearmanUmat(data)
}
\arguments{
\item{data}{the n by p raw data matrix}
}
\value{
a matrix with dimension p by p of spearman correlations
}
\description{
This routine calculates the SpearmanU, the pairwise covariance matrix estimator proposed in Oellererand Croux
}
