% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QUIC_CV.R
\name{cvQUIC}
\alias{cvQUIC}
\title{Cross validation to chose tuning parameter of QUIC}
\usage{
cvQUIC(
  data,
  k = 10,
  covest = cov,
  lambdas = seq(0.1, 1, 0.1),
  evaluation = negLLrobOmega,
  ...
)
}
\arguments{
\item{data}{The full dataset, should be a matrix}

\item{k}{number of folds}

\item{covest}{a *function* that takes a matrix to estimate covariance}

\item{lambdas}{a vector of tuning parameter to be tested}

\item{evaluation}{a *function* that takes only two arguments, the estimated *precision* and the test *covariace*, when NULL, we use negative log likelihood on test sets}

\item{...}{extra arguments send to QUIC}
}
\value{
a matrix with k rows, each row is the evaluation loss of that fold
}
\description{
This routine use k fold cross validation to chose tuning parameter
}
\examples{
cvQUIC(matrix(rnorm(100),20,5), msg = 0)
}
