% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nearPPSD}
\alias{nearPPSD}
\title{nearest positive semi-definite projection of a matrix}
\usage{
nearPPSD(X, eigenTol = 1e-06, convTol = 1e-07, psdTol = 1e-08, maxit = 1000L)
}
\arguments{
\item{X}{the matrix}

\item{eigenTol}{tolerance in eigen system, used in finding nearest positive matrix}

\item{convTol}{tolerance in cov, used in finding nearest positive matrix}

\item{psdTol}{tolerance in psd, used in finding nearest positive matrix}

\item{maxit}{max iterations in finding nearest positive matrix}
}
\value{
a matrix which is the nearest positive semi-definite matrix of input X
}
\description{
This routine calculate the nearest positive semi0definite projection
}
