% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Multivariate normal distribution with 0 mean}
\usage{
rmvnorm(n, Omega)
}
\arguments{
\item{n}{sample size}

\item{Omega}{**precision** matrix of dimension p by p}
}
\value{
a matrix with dimension n by p, each row is a sample
}
\description{
This routine samples multivarate normal distribution of mean 0 from precision matrix
}
