% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvt}
\alias{rmvt}
\title{Multivariate t distribution}
\usage{
rmvt(n, Omega, nu)
}
\arguments{
\item{n}{sample size}

\item{Omega}{**precision** matrix of dimension p by p}

\item{nu}{degree of freedom}
}
\value{
a matrix with dimension n by p, each row is a sample
}
\description{
This routine samples multivarate t distribution
}
