\name{RobustOmega-package}
\alias{RobustOmega-package}
\alias{RobustOmega}
\docType{package}
\title{
\packageTitle{RobustOmega}
}
\description{
\packageDescription{RobustOmega}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RobustOmega}
\packageIndices{RobustOmega}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RobustOmega}

Maintainer: \packageMaintainer{RobustOmega}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
