% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_CV.R
\name{negLLrobOmega}
\alias{negLLrobOmega}
\title{-log Likelihood on test set}
\usage{
negLLrobOmega(Sigma_hat, Sigma)
}
\arguments{
\item{Sigma_hat}{the estimated *covariance* matrix of training set}

\item{Sigma}{the *covariance* matrix of test sets}
}
\value{
-log likelihood
}
\description{
The default evaluation function in corss validation, -log liekihood on test set
}
