% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso_CV.R
\name{robglasso}
\alias{robglasso}
\title{glasso with robust covariance estimations}
\usage{
robglasso(
  data,
  covest = cov,
  rho = 0.1,
  CV = FALSE,
  k = 10,
  grids = 15,
  evaluation = negLLrobOmega,
  ...
)
}
\arguments{
\item{data}{raw data, shoule be a matrix}

\item{covest}{a *function* or string that takes a matrix to estimate covariance}

\item{rho}{a scalar or vector of tuning parameters, if CV=FALSE, shoule be a scalar, if CV=TRUE scalar input will be override and tuning parameter will be chosed based on CV}

\item{CV}{bool, whether doing corss validation for tuning parameter, if lambda is a scalar, the candidate will be chosen automatically by log spacing between 0.01 max covariance and max covariance with number of grids}

\item{k}{fold for corss validation if applicable}

\item{grids}{number of candidate tuning parameters in cross validation}

\item{evaluation}{a *function* or string that takes only two arguments, the estimated *precision* and the test *covariace*, when NULL, we use negative log likelihood on test sets}

\item{...}{extra argument sent to glasso::glasso}
}
\value{
a glasso return (see ?glasso::glasso), most important one is $X the estimated sparse precision,with an extra entry of tuning parameter lambda
}
\description{
This routine fits glasso using a robust covariance matrix
}
\examples{
robglasso(matrix(rnorm(100),20,5))
}
